Rename detector to evaluator
